% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckFT.R
\name{CheckFT}
\alias{CheckFT}
\title{Monitorización de la fase transversal de la depuración selectiva.}
\usage{
CheckFT(variables, units, edData, varParam, score, selInfo, invTrans.List,
  dTrans.List, lang = "EN")
}
\arguments{
\item{variables}{\code{vector} de tipo \code{character} con los nombres de las
variables para los que se realiza la monitorización.}

\item{units}{\linkS4class{data.table} con los identificadores de las unidades a
monitorizar.}

\item{edData}{objeto de clase \code{StQList} con los microdatos depurados de 
la encuesta utilizados para la construcción de los intervalos de validación
del periodo de referencia especificado en Periodo.}

\item{varParam}{objeto de clase \code{contObsPredModelParam} con información
de la fase transversal.}

\item{score}{\linkS4class{data.table} con el valor de la función score para
cada unidad en la muestra.}

\item{selInfo}{objeto de clase \code{StQ} con los datos del fichero FT obtenido
en la fase transversal del periodo de referencia especificado en Periodo.}

\item{invTrans.List}{\code{lista} con los nombres de las funciones inversas, si se ha hecho
transformación (en otro caso con el valor NULL), para cada uno de los edits.}

\item{dTrans.List}{\code{lista} con los nombres de las funciones derivadas, si se ha hecho
transformación (en otro caso con el valor NULL), para cada uno de los edits.}

\item{lang}{\code{vector} de tipo \code{character} y longitud uno con el 
idioma en que se quiere obtener la salida. Sus posibles valores son: 'EN'
(English) y 'SP' (Servicio Promotor) y su valor por defecto es 'EN'.}
}
\value{
La función devuelve una lista con dos componentes:
\item edTS: data.table con la serie de valores de las unidades para las 
variables especificadas.
\item Param: data.table con los parámetros utilizados en la fase transversal.
}
\description{
\code{CheckFT} obtiene los parámetros de la fase transversal de la 
depuración selectiva
}
\examples{
dontrun{
}

}
