% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckInterval.R
\name{CheckInterval}
\alias{CheckInterval}
\title{Monitorización de la fase longitudinal de la depuración selectiva.}
\usage{
CheckInterval(variables, units, edit.List, edits = names(edit.list), edData,
  rawData, edTS, intervalData, flagData, invTrans.List, dTrans.List,
  lang = "EN")
}
\arguments{
\item{variables}{\code{lista} con tres componentes: VarAnalisis, con los nombres 
de las variables implicadas en la contrucción de los controles de validación 
de la operación estadística considerada; VarAsociadas, con las variables 
que intervienen en el cálculo de cada variable de análisis y VarAuxiliares, con los
nombres de aquellas variables que quieran incluirse en la salida de la función.}

\item{units}{\linkS4class{data.table} con los identificadores de las unidades a
monitorizar.}

\item{edits}{\code{vector} de tipo \code{character} con los nombres de los
controles de validación que se quieren consultar.}

\item{edData}{objeto de clase \code{StQ} con los microdatos depurados de la
encuesta para el periodo consultado.}

\item{rawData}{objeto de clase \code{StQ} con los microdatos sin depurar de la 
encuesta para el periodo consultado.}

\item{edTS}{objeto de clase \code{StQ} con los microdatos depurados de la
encuesta utilizados para la construcción de los intervalos de validación del
periodo de referencia especificado en Periodo.}

\item{intervalData}{objeto de clase \code{StQ} con los datos del fichero FL obtenido
en la fase longitudinal del periodo de referencia especificado en Periodo.}

\item{flagData}{\code{lista} de \linkS4class{data.table} con la información
de los ficheros de errores.}

\item{invTrans.List}{\code{lista} con los nombres de las funciones inversas, si se ha hecho
transformación (en otro caso con el valor NULL), para cada uno de los edits.}

\item{dTrans.List}{\code{lista} con los nombres de las funciones derivadas, si se ha hecho
transformación (en otro caso con el valor NULL), para cada uno de los edits.}

\item{lang}{\code{vector} de tipo \code{character} y longitud uno con el 
idioma en que se quiere obtener la salida. Sus posibles valores son: 'EN'
(English) y 'SP' (Servicio Promotor) y su valor por defecto es 'EN'.}

\item{edit.list}{lista de vectores de tipo \code{character} con los nombres de
los controles de validación de la operación estadística considerada.}
}
\value{
La función devuelve una lista con dos componentes:
\item edTS: data.table con la serie de valores de las unidades para las 
variables implicadas en los edits.
\item edits: data.table con los intervalos de validación de cada Edit
consultado y los parámetros utilizados en su construcción: Pred (centro del
intervalo), Pred_error (error de predicción de la variable), HRFactor (factor
de amplitud) y Condicion (indicador de la aplicación o no del control.)
}
\description{
\code{CheckInterval} obtiene los parámetros de la fase longitudinal de la 
depuración selectiva
}
\examples{
dontrun{
Variables <- list(VarAnalisis = list(LCN = 'logv9' ,
                                     LEX = 'logvexis',
                                     LRCNPT = 'logRatioCNPT',
                                     LTACNCCAA = 'logitAngleCNCCAA',
                                     LTACNCNAE = 'logitAngleCNCNAE',
                                     LTAPTCCAA = 'logitAnglePTCCAA',
                                     LTACNPTCCAA = 'logitAngleCNPTCCAA'),
                  VarAsociadas = list(LCN = 'v9' ,
                                      LEX = 'vexis',
                                      LRCNPT = c('v9', 'vexis'),
                                      LTACNCCAA = c(vecCNCCAA, lagVecCNCCAA),
                                      LTACNCNAE = c(vecCNCNAE, lagVecCNCNAE),
                                      LTAPTCCAA = c(vecPTCCAA, lagVecPTCCAA),
                                      LTACNPTCCAA = c(vecCNCCAA, vecPTCCAA)),
                  VarAuxiliares = list(INCIDENC = 'incidenc'))
Suffixes <- c('W_1', 'W_3', 'noW_3')
Edit.List <- list(LCN = paste0('LCN_', Suffixes),
                  LEX = paste0('LEX_', Suffixes),
                  LRCNPT = paste0('LRCNPT_', Suffixes),
                  LTACNCCAA = paste0('LTACNCCAA_', Suffixes),
                  LTACNCNAE = paste0('LTACNCNAE_', Suffixes),
                  LTAPTCCAA = paste0('LTAPTCCAA_', Suffixes),
                  LTACNPTCCAA = paste0('LTACNPTCCAA_', Suffixes))  
                 
output <- CheckInterval(units = list(NOrden = '4602232599M'), 
                        edits = c('LCN','LTACNCNAE'), 
                        variables = Variables, 
                        edit.List = Edit.list, 
                        edData = FF.StQList[[Periodo]],
                        intervalData = FL.StQList)
}

}
