% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivL.R
\name{DerivL}
\alias{DerivL}
\title{Derivada de log(1+x) para cada uno de los elementos x del vector de entrada.}
\usage{
DerivL(p)
}
\arguments{
\item{p}{\code{Vector} de tipo \code{numeric} sobre el que se calculará la función.}
}
\value{
\code{Vector} de tipo \code{numeric} de longitud igual al vector de entrada \code{p} con
el valor que toma la función: 1 / (1 + x) para cada uno de los elementos del vector.
}
\description{
\code{DerivL} Calcula la derivada de log(1+x), es decir, 1 / (1 + x), para cada uno de los
elementos x del vector.
}
\examples{
dontrun{

}

}
