% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DerivLTA.R
\name{DerivLTA}
\alias{DerivLTA}
\title{Derivada de Logit(x) para cada uno de los elementos x del vector de entrada.}
\usage{
DerivLTA(p, adjust = 1e-15, unit = "degrees")
}
\arguments{
\item{p}{\code{Vector} de tipo \code{numeric} sobre el que se calculará la función.}

\item{adjust}{\code{Vector} de tipo \code{numeric} de longitud 1. Por defecto toma el valor
\code{10^-15}.}
}
\value{
\code{Vector} de tipo \code{numeric} de longitud igual al vector de entrada \code{p} con
el valor de la función derivada del Logit para cada uno de los elementos del vector.
}
\description{
\code{DerivLTA} Calcula la derivada de Logit(x) para cada uno de los elementos x del vector.
}
\examples{
dontrun{

}

}
