% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlagMatrix.R
\name{FlagMatrix}
\alias{FlagMatrix}
\title{Construcción de matrices indicadoras de controles activos}
\usage{
FlagMatrix(flagData, TypeEd = "P")
}
\arguments{
\item{flagData}{\linkS4class{data.table} con el fichero de controles activados 
a partir del cual se quiere obtener la tabla de salida.}

\item{TypeEd}{Vector \code{character} indicando si los controles han saltado
en primera ('P') o en última depuación ('U').}

\item{ed}{objeto de clase \linkS4class{StQ} del que se obtienen las unidades
para las que se quiere obtener la tabla de controles activos.}

\item{Edit.List}{lista los nombres de los controles cuya activación quiere 
consultarse.}

\item{by}{Vector \code{character} con los nombres de las variables de 
agrupación para el recuento de controles activos. Si no se especifica, se 
toma por defecto el valor del IDQual de los datos en el parámetro de entrada
ed.}
}
\value{
\code{\link{data.table}} con la matriz indicadora de activación de 
controles.
}
\description{
\code{FlagMatrix} devuelve una \code{\link{data.table}} con las variables 
binarias que indican para cada unidad estadística (fila) qué controles 
(columnas) se han activado.
}
\details{
Esta función lee el fichero FC cuyo nombre está especificado en 
\code{ed}, con diseño de registro especificado en el parámetro 
\code{Schema} y devuelve su contenido en una \code{data.table} con estructura
 matricial (unidades en filas y variables en columnas). Los nombres de las 
variables se asignan de acuerdo con el diseño de registro especificado.
}
\examples{
\dontrun{
data(FC)
}

}
