% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invL.R
\name{invL}
\alias{invL}
\title{Inversa de log(1+x) para cada uno de los elementos x del vector de entrada.}
\usage{
invL(p)
}
\arguments{
\item{p}{\code{Vector} de tipo \code{numeric} sobre el que se calculará la función.}
}
\value{
\code{Vector} de tipo \code{numeric} de longitud igual al vector de entrada \code{p} con
el valor que toma la función: exp(x) - 1 para cada uno de los elementos del vector.
}
\description{
\code{invL} Calcula la inversa de log(1+x), es decir, exp(x) - 1, para cada uno de los elementos
x del vector.
}
\examples{
dontrun{

}

}
