% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validate.R
\name{Validate}
\alias{Validate}
\title{Validación de la fase longitudinal de la depuración selectiva.}
\usage{
Validate(Data, intervalData, variables, lang = "EN")
}
\arguments{
\item{Data}{objeto de clase \code{StQ} con los microdatos de la encuesta para los que se quiere
realizar la consulta.}

\item{intervalData}{objeto de clase \code{StQ} con los datos del fichero FL obtenido en la fase
longitudinal del periodo al que pertenecen los datos del parámetro \code{Data}.}

\item{variables}{\code{lista} con tres componentes: VarAnalisis, con los nombres de las variables
implicadas en la contrucción de los controles de validación de la operación estadística
considerada; VarAsociadas, con las variables que intervienen en el cálculo de cada variable de
análisis y VarAuxiliares, con los nombres de aquellas variables que quieran incluirse en la
salida de la función.}

\item{lang}{\code{vector} de tipo \code{character} y longitud uno con el idioma en que se quiere
obtener la salida. Sus posibles valores son: 'EN'(English) y 'SP' (Servicio Promotor) y su valor
por defecto es 'EN'.}
}
\value{
La función devuelve una \code{data.table} con los intervalos de validación para cada
control de cada variable de análisis, el valor informado de ésta y TRUE o FALSE en la columna
Flag dependiendo de si el control debería haberse activado o no, así como otra información
complementaria.
}
\description{
\code{Validate} comprueba si debería haber saltado o no cada control, en función de los valores
informados y los intervalos obtenidos para cada control en la fase longitudinal de la depuración
selectiva.
}
\examples{
dontrun{
Variables.list <- list(VarAnalisis = list(LCN = 'logv9' ,
                                          LEX = 'logvexis',
                                          LRCNPT = 'logRatioCNPT',
                                          LTACNCCAA = 'logitAngleCNCCAA',
                                          LTACNCNAE = 'logitAngleCNCNAE',
                                          LTAPTCCAA = 'logitAnglePTCCAA',
                                          LTACNPTCCAA = 'logitAngleCNPTCCAA'),
                      VarAsociadas = list(LCN = 'v9' ,
                                          LEX = 'vexis',
                                          LRCNPT = c('v9', 'vexis'),
                                          LTACNCCAA = c(vecCNCCAA, lagVecCNCCAA),
                                          LTACNCNAE = c(vecCNCNAE, lagVecCNCNAE),
                                          LTAPTCCAA = c(vecPTCCAA, lagVecPTCCAA),
                                          LTACNPTCCAA = c(vecCNCCAA, vecPTCCAA)),
                      VarAuxiliares = list(INCIDENC = 'incidenc'))
                 
output <- Validate(Data = FF, 
                   intervalData = FL,
                   variables = Variables.list,
                   lang = 'SP')
}

}
